Initial changes to remove hardcoded paths
The task is to make the backup module work on non pi systems.

I have modified this so that the config file now resides in the php folder and is read by both the php side and the bash side.  I am not sure if this is a security issue or not.

I have also added in some file exist checks for non EmonPi setups.

Phpfiwa was not backed up.

This needs more development in identifying different base systems and reacting accordingly.

feedwriter is not stopped currently (commented out) as I dot have it installed.  The service could be checked for existing & running.

As I say, just a start on this.